% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeAPRGRG.R
\name{codeAPRDRG}
\alias{codeAPRDRG}
\title{APR DRG coder}
\usage{
codeAPRDRG(mydata, Severity = NULL)
}
\arguments{
\item{mydata}{A vector of APR DRG Numners or Names}

\item{Severity}{an optional vector of integer APR serverity, same length of mydata}
}
\value{
a data frame with aprDrgWeight, ALOS and CostOutThresh
}
\description{
recodes an APRDRG to APR wieght, mean Length of Stay and Cost outlier thresh
}
\details{
APR DRG 2014 table
}
\author{
Matthew Davis
}
