% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeMSDRG.R
\name{codeMSDRG}
\alias{codeMSDRG}
\title{MSDRG to LOS and Weight}
\usage{
codeMSDRG(mydata)
}
\arguments{
\item{mydata}{a vector of MSDRG Names or MSDRG Numbers}
}
\value{
a data frame with msDRG weight, GMLOS, and AMLOS
}
\description{
recodes an msDRG to the CMS-1655-P-FY-2017-TABLE-5-_Weight_File_FR
}
\details{
GMLOS = geometric length of stay, AMLOS arithemtic length of stay
}
\author{
Matthew Davis
}
