% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooksDistanceOptimizer.R
\name{cooksDistanceOptimizer}
\alias{cooksDistanceOptimizer}
\title{Test models in an enviroment}
\usage{
cooksDistanceOptimizer(fit, nfits = 10, p = 0.5)
}
\arguments{
\item{fit}{a glm model}

\item{nfits}{number of models to fit}

\item{p}{probablity cut off if binomial is use used}
}
\value{
a dataframe of performance metrics based on removing influential points
}
\description{
This function finds the optimal number of influential points to remove from a linear model
}
\details{
the CooksDistance Cut is the point at which to remove any points with greater distance
}
\author{
Matthew Davis
}

