% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrameFactDiffs.R
\name{dataFrameFactDiffs}
\alias{dataFrameFactDiffs}
\title{data frame Factor Differences}
\usage{
dataFrameFactDiffs(df1, df2, verbose = FALSE)
}
\arguments{
\item{df1}{a data frame}

\item{df2}{a data frame to compair with}

\item{verbose}{print debugging output}
}
\value{
a data frame with pvals from McNemars test, xMean, yMean is the % that a level appears
}
\description{
takes two data frames and finds differences in the factor distrobutions
}
