% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorDiffs.R
\name{factorDiffs}
\alias{factorDiffs}
\title{factor Differences}
\usage{
factorDiffs(x, y, reOrder = TRUE, verbose = FALSE)
}
\arguments{
\item{x}{a factor vector}

\item{y}{a factor vector}

\item{reOrder}{whether to re order by the p value}

\item{verbose}{print debugging output}
}
\value{
a vector of p values for each factor levels
}
\description{
takes two factor vectors and preforms McNemar test and returns a p value for every factor level
}
