% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDfDistoDiffs.R
\name{find2dfDiffs}
\alias{find2dfDiffs}
\title{Find the Optimum Kmeans Cluster Model}
\usage{
find2dfDiffs(df1, df2, verbose = TRUE)
}
\arguments{
\item{df1}{a data frame}

\item{df2}{a data frame}

\item{verbose}{print debugging output}
}
\value{
a data frame of ss errors by algorythim types
}
\description{
builds a bunch of cluster models to find the optimum number of clusters
}
\details{
This predict UHC Work RVUS from Work RVUs
}
\author{
Matthew Davis
}
