% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBestAutoencoder.R
\name{findBestAutoEncoder}
\alias{findBestAutoEncoder}
\title{Find the Best single hidden layer Autoencoder}
\usage{
findBestAutoEncoder(trainData, maxHiddenLayers = 1, maxNodes = 2,
  n.models = 40, rescale = TRUE, max.int = 1000)
}
\arguments{
\item{trainData}{a data frame or matrix with numeric columns}

\item{maxHiddenLayers}{currently not implimented}

\item{maxNodes}{max number of nodes in hidden layer}

\item{n.models}{number of models to build}

\item{rescale}{logical rescales the data}

\item{max.int}{max number of iterations}
}
\value{
\itemize{
     \item model is the best fit model i
     \item a.grid the grid used to build models that were compaired
     \item sse vector of sum squared erros
     \item BestParams df of the best tuning parameters for the auto encoder
     \item colNames used to for the autoencoder (any character cols are filtered out)}
}
\description{
this creates a number of autoencoders and compairs sse to find optimum
}
\details{
finds the best single hidden layer autoencoder
}
\author{
Matthew Davis
}
