% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClusters.R
\name{findClusters}
\alias{findClusters}
\title{Find the Optimum Kmeans Cluster Model}
\usage{
findClusters(mydata, n = 10, seed = 2012, plotIt = TRUE)
}
\arguments{
\item{mydata}{list of models}

\item{n}{a data frame to test the model against}

\item{seed}{set seed for randomization}

\item{plotIt}{whether to plot solution}
}
\value{
a data frame of ss errors by algorythim types
}
\description{
builds a bunch of cluster models to find the optimum number of clusters
}
\details{
This predict UHC Work RVUS from Work RVUs
}
\author{
Matthew Davis
}
