% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTestenv.R
\name{fitTest.env}
\alias{fitTest.env}
\title{Test models in an enviroment}
\usage{
fitTest.env(model.env, testSet = toTest, treatment = NULL)
}
\arguments{
\item{model.env}{enviroment where 'train' class models are stroed}

\item{testSet}{a data frame to test the model against}

\item{treatment}{a 'teatment' class object from vtreat package (if this was used as a pre proc step on input data)}
}
\value{
vector or performance metrics for the model
}
\description{
usefull if you save a bunch models in an enviroment and want to test them all
}
\details{
This predict UHC Work RVUS from Work RVUs
}
\author{
Matthew Davis
}
