% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTest.R
\name{fitTest}
\alias{fitTest}
\title{Test a set of models in an enviroment}
\usage{
fitTest(model, testSet, Treatment = NULL, className = NULL, p = 1)
}
\arguments{
\item{model}{a model of class 'train' from caret}

\item{testSet}{a data frame to test the model against}

\item{Treatment}{a 'teatment' class object from vtreat package (if this was used as a pre proc step on input data)}

\item{className}{optional class Name (target Class), if not provided, uses colnames(testSet)-coefNames}

\item{p}{psign (Vtreat parameter that trims off insigicant varibles, set to 1 for no trimming)}
}
\value{
vector or performance metrics for the model
}
\description{
tests a model with a vtreatment and a test set
}
\details{
This predict UHC Work RVUS from Work RVUs
}
\author{
Matthew Davis
}
