% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullPreProcess.R
\name{fullPreProcess}
\alias{fullPreProcess}
\title{full Pre Process}
\usage{
fullPreProcess(mydata, dmv = NULL, pp = NULL, treat = NULL)
}
\arguments{
\item{mydata}{a data frame to pre process}

\item{dmv}{dummies from the dummyVars function}

\item{pp}{preProces from the caret::preProcess}

\item{treat}{treatment from vtreat designTreatment}
}
\value{
a data frame of preprocessed data
}
\description{
looks up zip codes 2014 IRS tax data aggregated
}
\details{
for use to create a preprocessed dataset for predictive modeling
}
\author{
Matthew Davis
}
