% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getZip.R
\name{getZip}
\alias{getZip}
\title{get Zip}
\usage{
getZip(zipCodes, pca = FALSE)
}
\arguments{
\item{zipCodes}{a vector of vip codes}

\item{pca}{logical, whether or not to return principal components}
}
\description{
looks up zip codes 2014 IRS tax data aggregated
}
\details{
returns a dataframe with demographics for each zipcode,
If the zipcode is not found, it returns the mean of all zipcodes
}
\author{
Matthew Davis
}
