% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs.R
\docType{data}
\name{irs}
\alias{irs}
\title{IRS 2014 Data
IRS 2014 Data based on zipcodes}
\format{An object of class \code{data.frame} with 27781 rows and 19 columns.}
\source{
https://www.irs.gov/uac/soi-tax-stats-individual-income-tax-statistics-2014-zip-code-data-soi
data set from zipcode lookup purposes.
}
\usage{
irs
}
\description{
IRS 2014 Data
IRS 2014 Data based on zipcodes
}
\keyword{datasets}
