% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnCharacterEncoding.R
\name{learnCharacterEncoding}
\alias{learnCharacterEncoding}
\title{Lean Character Encoding}
\usage{
learnCharacterEncoding(mydata, factCols = NULL, maxLevs = 50, minFreq = 1,
  verbose = FALSE)
}
\arguments{
\item{mydata}{a data frame containing factors}

\item{factCols}{a character of columns in mydata}

\item{maxLevs}{maximum number of factor levels to encode}

\item{minFreq}{minimum frequency to spawn an encoded col}

\item{verbose}{pring debugging}
}
