% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelMatcher.R
\name{levelMatcher}
\alias{levelMatcher}
\title{Level Matcher}
\usage{
levelMatcher(dmv = NULL, newData, oldData = NULL, verbose = TRUE)
}
\arguments{
\item{newData}{data frame}

\item{oldData}{data frame}

\item{verbose}{print debugging output}
}
\value{
new data frame, where NA is used when an unknow level is present
}
\description{
applies one set of levels from a data frame to another data frame
}
