% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticDifference.R
\name{logisticDifference}
\alias{logisticDifference}
\title{Logistic Difference}
\usage{
logisticDifference(model, testOb, mydata = NULL, refOb = NULL,
  verbose = FALSE)
}
\arguments{
\item{model}{a model of class glm, rpart or ranger}

\item{testOb}{a data frame with one row to predict with the model}

\item{mydata}{a data frame with columns included in model}

\item{refOb}{a data frame with one row for a reference observation}

\item{verbose}{print debugging output}
}
\value{
a vector of estimated absolute influences
}
\description{
a function that interates through model predictors to estimate impact
}
\author{
Matthew Davis
}
