% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mostLikelyFact.R
\name{mostLikelyFact}
\alias{mostLikelyFact}
\title{Most Likely Factor}
\usage{
mostLikelyFact(factorVar, nFacts = 1)
}
\arguments{
\item{factorVar}{a factor variable}

\item{nFacts}{number of of factors to return}
}
\value{
a vector of frequencies of the most frequent Classes
}
\description{
finds the most frequent factor, or if number finds the mode
}
\author{
Matthew Davis
}
