% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mostLikelyOb.R
\name{mostLikelyOb}
\alias{mostLikelyOb}
\title{Most Likely Factor}
\usage{
mostLikelyOb(mydata, useMedian = TRUE)
}
\arguments{
\item{mydata}{a data frame}

\item{useMedian}{calc median for numerics, if FALSE calculates mode}
}
\value{
a vector of frequencies of the most frequent Classes
}
\description{
finds the most frequent factor, or if number finds the mode
}
\author{
Matthew Davis
}
