% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiColSearch.R
\name{multiColSearch}
\alias{multiColSearch}
\title{Multiple Column Search}
\usage{
multiColSearch(mydata, val, cols = NULL, returnMat = FALSE, all = TRUE)
}
\arguments{
\item{mydata}{a data frame}

\item{val}{a character vector of values to search for in columns}

\item{cols}{a character vector of column names in mydata to search}

\item{returnMat}{logical return a matrix of logical columns}

\item{all}{logical whether to return TRUE for all matches , if FALSE than any Matchs (ignored when returnMat)}
}
\value{
a  logical vector or matrix of logical vectors
}
\description{
searches for values in multiple columns
}
\details{
This predict UHC Work RVUS from Work RVUs
}
\author{
Matthew Davis
}
