% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAtoUnknownFunctions.R
\name{NAtoValueDf}
\alias{NAtoValueDf}
\title{NA To Value for Dataframe}
\usage{
NAtoValueDf(mydata, numCols = NULL, factCols = NULL, numVal = NULL,
  newLevName = "unKnown")
}
\arguments{
\item{mydata}{a dataframe}

\item{numCols}{a char vector of numeric column names to apply the na to zero}

\item{factCols}{a char vector of factor column names}

\item{numVal}{a numeric value to replace NA in numeric columns}

\item{newLevName}{new level Name for NA levels}
}
\value{
the mydata data frame with num Nas mapped to numVal and fact Nas mapped to newLevName
}
\description{
maps all numeric or only specified numeric columns in a data frame to val
}
\details{
if numVal or newLevName is null, then no action is taken that type of column
}
