% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticDifferenceDf.R
\name{plot.logisticDiff}
\alias{plot.logisticDiff}
\title{Plot method for class logisticDiff}
\usage{
\method{plot}{logisticDiff}(obj, n.obs = 50, n.vars = 10,
  indvidualPlot = FALSE, bins = 50, verbose = FALSE)
}
\arguments{
\item{obj}{an object of class logisticDiff}

\item{n.obs}{number of obs to be used}

\item{indvidualPlot}{to individual observation influence}

\item{bins}{number of bins on histogram plot}

\item{verbose}{print debugging}
}
\description{
A plot method of variable inflence objects
}
