% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preformanceByCatagory.R
\name{plot.preformByCat}
\alias{plot.preformByCat}
\title{Plot method for preformance by category data frame}
\usage{
\method{plot}{preformByCat}(obj, p = 1, verbose = FALSE, labLen = 20)
}
\arguments{
\item{obj}{data frame output from preformanceByCatagory}

\item{p}{integet show plot 1 or 2 (different preformance metric)}

\item{verbose}{print output}
}
\description{
Treemap plot of performance by catagory, size is num of obs with that cat
}
