% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotImp.R
\name{plotImp}
\alias{plotImp}
\title{plotImp}
\usage{
plotImp(fit, plotIt = TRUE)
}
\arguments{
\item{fit}{a genaric model that caret varImp works on}

\item{plotIt}{Whether or not to plot the varImp}
}
\description{
looks up zip codes 2014 IRS tax data aggregated
}
\details{
returns a dataframe with demographics for each zipcode,
If the zipcode is not found, it returns the mean of all zipcodes
}
\author{
Matthew Davis
}
