% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnCharacterEncoding.R
\name{predict.catagoryEncoder}
\alias{predict.catagoryEncoder}
\title{predict method for catagorical encoder object}
\usage{
\method{predict}{catagoryEncoder}(catEncoder, newData, sparse = TRUE)
}
\arguments{
\item{catEncoder}{catagoryEncoder object}

\item{newData}{a new Data frame}

\item{sparse}{return a sparse matrix}
}
