% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictkmeans.R
\name{predict.kmeans}
\alias{predict.kmeans}
\title{predict method for Kmeans}
\usage{
\method{predict}{kmeans}(fit, newData)
}
\arguments{
\item{fit}{kmeans class object}

\item{newData}{data frame of new data}
}
\value{
a data frame of ss errors per cluster, cluster classification and totalSSE
}
\description{
builds a bunch of cluster models to find the optimum number of clusters
}
\details{
cluster reported back as a factor in the output data frame
}
\author{
Matthew Davis
}
