% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictkmeans.R
\name{predictKmeans}
\alias{predictKmeans}
\title{predict method for Kmeans}
\usage{
predictKmeans(fit, newdata, verbose = FALSE)
}
\arguments{
\item{fit}{kmeans model}

\item{newdata}{data frame with new data to assign clusters to}

\item{verbose}{print debugging output}
}
\value{
a data frame of ss errors per cluster, cluster classification and totalSSE
}
\description{
builds a bunch of cluster models to find the optimum number of clusters
}
\details{
cluster reported back as a factor in the output data frame
}
\author{
Matthew Davis
}
