% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictvtreat.R
\name{predictVtreat}
\alias{predictVtreat}
\title{Predict with vtreat}
\usage{
predictVtreat(fit, mydata, treatment = NULL, p = 1)
}
\arguments{
\item{fit}{a model of class 'train' from caret}

\item{mydata}{a data frame to test the model against}

\item{treatment}{a 'teatment' class object from vtreat package (if this was used as a pre proc step on input data)}

\item{p}{psign (Vtreat parameter that trims off insigicant varibles, set to 1 for no trimming)}
}
\value{
vector or performance metrics for the model
}
\description{
this creates a prediction method with a treatment using vtreat
}
\details{
This predict UHC Work RVUS from Work RVUs
}
\author{
Matthew Davis
}

