% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preformanceByCatagory.R
\name{preformanceByCatagory}
\alias{preformanceByCatagory}
\title{Preformance Test By Catagory}
\usage{
preformanceByCatagory(preds, obs, cats, classify = FALSE, minFreq = 10,
  verbose = FALSE)
}
\arguments{
\item{preds}{a vector of predictions}

\item{obs}{a vector of observations}

\item{cats}{a vector of catagories}

\item{classify}{logical use classification}

\item{minFreq}{minium frequency of classes}

\item{verbose}{whether to print debugging output}
}
\description{
test model preformance by a catagory
}
