% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reOrderDataFrameFactors.R
\name{reOrderAll}
\alias{reOrderAll}
\title{Re Order Data Frame Factors}
\usage{
reOrderAll(mydata, factCols = NULL)
}
\arguments{
\item{mydata}{a data frame}

\item{factCols}{a character vector of factor columns (optional)}
}
\value{
a data frame with factors levels so most frequent is first

a data frame or factor with the the most frequent level first for each factor
}
\description{
this function takes a data frame and returns it with reordered levels
}
\details{
only reoders factor levels
}
\author{
Matthew Davis
}
