% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reOrderDataFrameFactors.R
\name{reOrderDataFrameFactors}
\alias{reOrderDataFrameFactors}
\title{Re Order Data Frame Factors}
\usage{
reOrderDataFrameFactors(mydata)
}
\arguments{
\item{mydata}{a data frame}
}
\value{
a data frame with factors levels so most frequent is first
}
\description{
this function takes a data frame and returns it with reordered levels
}
\details{
only reoders factor levels
}
\author{
James Zeng
}
