% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{size}
\alias{size}
\title{size}
\usage{
size(mydata, units = "Mb")
}
\arguments{
\item{mydata}{an r object}

\item{units}{units in the format(object.size) method}
}
\value{
a data frame of preformance metrics for all the models in the enviroments
}
\description{
usefull if you save a bunch models in an enviroment and want to test them all
}
\details{
this function is for choosing the best model based on a test set
}
\author{
Matthew Davis
}
