% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTestenv.R
\name{testEnv2class}
\alias{testEnv2class}
\title{Test Enviroment 2 class}
\usage{
testEnv2class(mydata, e, classCol = "Class", predVars = NULL,
  levs = c("notReAdmitted", "reAdmitted"), referenceLevNum = 2)
}
\arguments{
\item{mydata}{data frame to use as test data}

\item{e}{enviroment with models of class train are stored}

\item{classCol}{name of the class column}

\item{predVars}{column of predictor varaibles (optional), guess of the first model if NULL}

\item{levs}{levels of the class column}

\item{referenceLevNum}{of the reference level to consider success (ie typically 1 from 0/1 class)}
}
\value{
a data frame of preformance metrics for all the models in the enviroments
}
\description{
usefull if you save a bunch models in an enviroment and want to test them all
}
\details{
this function is for choosing the best model based on a test set
}
\author{
Matthew Davis
}
