% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniformNoiseAdd.R
\name{uniformNoiseAdd}
\alias{uniformNoiseAdd}
\title{Unifrom Noise Adder}
\usage{
uniformNoiseAdd(mydata, percent = 0.1, seed = 2016)
}
\arguments{
\item{mydata}{a dataframe or vector}

\item{percent}{pecent of noise to add from 0 to 1}

\item{seed}{seed for randomization}
}
\value{
vector or performance metrics for the model
}
\description{
used for adding an amount of noise to a data frame
}
\details{
uses min*percet and max*percent to run runif
}
\author{
Matthew Davis
}
