% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeZipcode.R
\name{zipCodeDems}
\alias{zipCodeDems}
\title{Find the Best single hidden layer Autoencoder}
\usage{
zipCodeDems(zipcodes, pca = FALSE)
}
\arguments{
\item{zipcodes}{a vector of zipcodes}

\item{pca}{logical, to return principal components of IRS data}
}
\description{
looks up zip codes 2014 IRS tax data aggregated
}
\details{
\itemize{
 \item AGI 1 = $1 under $25,000 :
 \item AGI 2 = $25,000 under $50,000:
 \item AGI 3 = $50,000 under $75,000:
 \item AGI 4 = $75,000 under $100,000:
 \item AGI  5 = $100,000 under $200,000:
 \item AGI  6 = $200,000 or more}
}
\author{
Matthew Davis
}
