% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticDifferenceDf.R
\name{logisticDifferenceDf}
\alias{logisticDifferenceDf}
\title{Logistic Difference for Data Frames}
\usage{
logisticDifferenceDf(model, testObs, refOb = NULL, mydata = NULL,
  verbose = TRUE)
}
\arguments{
\item{model}{a model of class glm, rpart or ranger}

\item{testObs}{a data frame with one row to predict with the model}

\item{refOb}{a data frame with one row for a reference observation}

\item{mydata}{a data frame with columns included in model}

\item{verbose}{print debugging output}
}
\value{
a vector of estimated absolute influences
}
\description{
a function that interates through model predictors to estimate impact
}
\author{
Matthew Davis
}
