% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multi Plot Function}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{plotlist}{a list of ggplot plots}

\item{cols}{number of columns}

\item{layout}{plot layout}
}
\value{
multiple ggplots combined as one
}
\description{
combines multiple ggplots
}
\author{
cookbook-R
}
