% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usableColumns.R
\name{usableColumns}
\alias{usableColumns}
\title{Usable Columns}
\usage{
usableColumns(mydata, maxUnique = 100, maxPrecentOneClass = 0.99,
  verbose = FALSE)
}
\arguments{
\item{mydata}{a data frame}

\item{maxUnique}{maxim number of unique values in a factor column}

\item{verbose}{print troubleshouting information}

\item{seed}{max Percentage of the majority class}
}
\value{
a vector of column names
}
\description{
returns a vector of columns names usable for machine learning models
}
\details{
This predict UHC Work RVUS from Work RVUs
}
\author{
Matthew Davis
}
